Move CodeExtension from TwigBridge to WebProfilerBundle
