+++ src/Symfony/Bundle/TwigBundle/Resources/config/twig.php
@@ -17,7 +17,6 @@ use Symfony\Bridge\Twig\DataCollector\TwigDataCollector;
 use Symfony\Bridge\Twig\ErrorRenderer\TwigErrorRenderer;
 use Symfony\Bridge\Twig\EventListener\TemplateAttributeListener;
 use Symfony\Bridge\Twig\Extension\AssetExtension;
-use Symfony\Bridge\Twig\Extension\CodeExtension;
 use Symfony\Bridge\Twig\Extension\ExpressionExtension;
 use Symfony\Bridge\Twig\Extension\HtmlSanitizerExtension;
 use Symfony\Bridge\Twig\Extension\HttpFoundationExtension;
@@ -105,10 +104,6 @@ return static function (ContainerConfigurator $container) {
 
         ->set('twig.extension.assets', AssetExtension::class)
             ->args([service('assets.packages')])
-
-        ->set('twig.extension.code', CodeExtension::class)
-            ->args([service('debug.file_link_formatter')->ignoreOnInvalid(), param('kernel.project_dir'), param('kernel.charset')])
-            ->tag('twig.extension')
 
         ->set('twig.extension.routing', RoutingExtension::class)
             ->args([service('router')])
